Save the column index to allow it to be printed for errors.
This one assumes a \t char equals 4 spaces since i don't know
how to make it dynamic yet. A few tests seem to indicate that the
column index is at the right place whenever an error comes up,
as long as there's not more than a single \t. Not sure why it
doesn't get the following ones yet though...