General fixes and cleanups to the compiler.
There were a few small issues remaining on the features that were
implemented a while ago. Most of those were fixed now (as can be
seen by the test files being parsed successfully when some of them
failed before even with the features that they need being
implemented). Some files still fail due to missing features, and
also because some of them are meant to fail as well.