Rework error reporting.
This fixes #3 somewhat. About time we nuked this one, even though
the error reporting isn't as fully fledged as GCC or other C
compilers. Instead of reporting the column, we report the token
number instead. It also works with the "panic mode" error
reporting, where a single error forces the compiler to stop
analyzing the rest of the code.